[iOS] Set content length if not provided by server, and using "?" to get suggested filename during download
Use Ranges header or Known-Content-Length special header in the Task to be used if the server does not provide Content-Length header. This enables progress updates for downloads of known size, even if the server does not provide content length.

"?" as filename will trigger parsing of response header for Content-Disposition, or last path segment of URL