Refactor, moving away from using `awaitGroup` and moving the `enqueueAndAwait` logic to the `BaseDownloader`.
Instead, now use `awaitTasks` to determine if an update is related to a convenience function.

This also makes it possible to remove references to `modifiedTasks` in the `PersistentStorage` interface and everywhere else where this is called.

Improved the logic for replacing the task's `updates` field appropriate to the way it is used in a convenience download (e.g. enforce progress updates if a progress callback is supplied). We still allow incorrect `updates` for backward compatibility, but now show an explicit warning. As long as the app stays alive, pause/resume will use the task with modified updates field, but this is not guaranteed, and this is why we want the developer to set the updates field properly