[iOS] Fix for issue #420 related to task count in group notifications.
Modifies the logic for group notifications such that the numTotal count increases when a task that is
part of a group notification is enqueued (as opposed to
when the task starts running, as is the case for
tasks not part of a group notification).
This ensures proper count especially when used together with a holding queue.

Also fixes a bug where the 'handleEventsForBackgroundURLSession' method was not called, because it was associated with the UrlSessionDelegate instead of the UIApplicationDelegate forwarding included with the FlutterPlugin protocol (that BDPlugin conforms to).