[Android] Fix issue #415 by expanding the type of Android URI that can be used to upload a file.
 Was MediaStore URIs only, now accepts any Android URI, e.g. one provided by a document provider such as a file picker.

 If the URI cannot provide a file size, the upload will fail.

Modifies the logic for group notifications such that the numTotal count increases when a task that is
part of a group notification is enqueued (as opposed to
when the task starts running, as is the case for
tasks not part of a group notification).
This ensures proper count especially when used together with a holding queue.

Also fixes a bug where the 'handleEventsForBackgroundURLSession' method was not called, because it was associated with the UrlSessionDelegate instead of the UIApplicationDelegate forwarding included with the FlutterPlugin protocol (that BDPlugin conforms to).