(fix) skip test 'One high priority task among regular ones' because it is flaky
Required 20 ms elapsed time after high priority download, but changed that to 5 to reduce likelihood that the test fails. It still sometimes fails.

Note this also reduces the usefulness of the test, but it's hard to test on fast networks.

As a result, I have now set 'skip: true' as there is very little change in this code. Can be reverted if work is done that affects priorities